refactor(scheduler): address bot review on cron status sync
- Move status-file sync into adapters/nonebot/cron_status_sync.py; scheduler_plugin only registers it

- Expose updated_at to frontend; show snapshot time and stale hint when snapshot is older than 5 minutes

- Harden status file reader against valid non-dict JSON; symmetric updated_at key in fallback responses

- Isolate _job_run_results in tests; drop _STATUS_PATH aliases in favor of the paths.py constant
